requirements.txt
streamlit>=1.30.0
plotly>=5.18.0
pandas>=2.1.0
PyMuPDF>=1.23.0
google-generativeai>=0.3.0
groq>=0.4.0
Pillow>=10.0.0
